phân quyền

+++ app/Http/Controllers/admin/Admin_groupController.php
@@ -12,8 +12,10 @@ class Admin_groupController extends Controller
     public function __construct(){
         $this->middleware('auth');
     }
-    public function getList()
+    public function getList(Request $request)
     {
+        if(check_permision($request->session()->get('data_session'),7,'_view') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $admin_group = Admin_groupModel::all();
         return view('admin.admin_group.index',
             [   'title'=>'Nhóm Quản trị Admin',
@@ -22,11 +24,15 @@ class Admin_groupController extends Controller
         );
            
     }
-    public function getAdd()
+    public function getAdd(Request $request)
     {
+        if(check_permision($request->session()->get('data_session'),7,'_add') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         return view('admin.admin_group.add');    
     }
     public function postAdd(Request $request){
+        if(check_permision($request->session()->get('data_session'),7,'_add') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $this->validate($request,
             [
                 'name' => 'required|unique:admin_group,name|min:3|max:255'
@@ -43,11 +49,15 @@ class Admin_groupController extends Controller
         $admin_group->save();
         return redirect('admin/admin_group/add')->with('thongbao','Thêm thành công');
     }
-    public function getEdit($id){
+    public function getEdit(Request $request,$id){
+        if(check_permision($request->session()->get('data_session'),7,'_edit') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $admin_group = Admin_groupModel::find($id);
         return view('admin.admin_group.edit',['data' => $admin_group]);    
     }
     public function postEdit(Request $request,$id){
+        if(check_permision($request->session()->get('data_session'),7,'_edit') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $admin_group = Admin_groupModel::find($id);
         $this->validate($request,
             [
@@ -64,8 +74,10 @@ class Admin_groupController extends Controller
         $admin_group->save();
         return redirect('admin/admin_group/edit/'.$id)->with('thongbao','Cập nhật thành công');    
     }
-    public function getDelete($id)
+    public function getDelete(Request $request,$id)
     {
+        if(check_permision($request->session()->get('data_session'),7,'_delete') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $admin_group = Admin_groupModel::find($id);
         $admin_group->delete();
         return redirect('admin/admin_group/list')->with('thongbao','Xóa thành công id:' .$id);
