+++ app/Http/Controllers/admin/UsersController.php
@@ -15,9 +15,10 @@ class UsersController extends Controller
     public function __construct(){
         $this->middleware('auth');
     }
-    public function getList()
+    public function getList(Request $request)
     {
-
+        if(check_permision($request->session()->get('data_session'),7,'_view') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $users = UsersModel::orderBy('id','DESC')->get();
         return view('admin.users.index',
             [   'title'=>'Thành viên',
@@ -26,12 +27,16 @@ class UsersController extends Controller
         );
            
     }
-    public function getAdd()
+    public function getAdd(Request $request)
     {
+        if(check_permision($request->session()->get('data_session'),7,'_add') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $group = Admin_groupModel::all();
         return view('admin.users.add',['group' => $group]);    
     }
     public function postAdd(Request $request){
+        if(check_permision($request->session()->get('data_session'),7,'_add') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
        $this->validate($request,
             [
                 'name' => 'required|min:3|max:255',
@@ -77,13 +82,16 @@ class UsersController extends Controller
         $users->save();
         return redirect('admin/users/add')->with('thongbao','Thêm thành công');
     }
-    public function getEdit($id){
+    public function getEdit(Request $request,$id){
+        if(check_permision($request->session()->get('data_session'),7,'_edit') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $users = UsersModel::find($id);
         $group = Admin_groupModel::all();
         return view('admin.users.edit',['data' => $users,'group'=>$group]);    
     }
     public function postEdit(Request $request,$id){
-        
+        if(check_permision($request->session()->get('data_session'),7,'_edit') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $this->validate($request,
             [
                 'name' => 'required|min:3|max:255',
@@ -126,8 +134,10 @@ class UsersController extends Controller
         $users->save();
         return redirect('admin/users/edit/'.$id)->with('thongbao','Cập nhật thành công');    
     }
-    public function getDelete($id)
+    public function getDelete(Request $request,$id)
     {
+        if(check_permision($request->session()->get('data_session'),7,'_delete') != 1)
+            return redirect('admin/permision')->with('thongbao','Bạn không có quyền thực hiện chức năng này');
         $users = UsersModel::find($id);
         $users->delete();
         return redirect('admin/users/list')->with('thongbao','Xóa thành công id:' .$id);
